Move page components to pages, remove proptypes
