+++ src/constants/links.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   github: 'https://github.com/5-gwoap/adopt-a-family',
   email: 'mailto:5gwoap@gmail.com',
   license: 'https://opensource.org/licenses/mit-license.html',
