+++ src/pages/Pairing.jsx
