+++ src/pages/Registration.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Form from './Form';
+import Form from '../components/Form';
 
 const Registration = (props) => (
   <div>
